Complete exercise
Note: Tried different flex inputs for the cardbox as I thought
the sidebar would grow vertically to match the cardbox's height.
Either it doesn't work this way, or something else prevented the
cardbox from growing. I'll have to figure that part out later.

I understand from the solution that it was the wrapper which needed
flex: 1, and I see how that makes sense. I'm going to leave my bootleg
solution in placee (which was to add an empty white box at the bottom
of the cards to get the sidebar to touch the footer when the cards
didn't reach the bottom of the screen) for posterity.

It left a small gap below the cards when you had enough to create a
scrollbar, but A) There are no explicit instructions that this is
unacceptable, B) As I re-read the hints, I realize that I misunderstood
them and that's how I came up with the idea lol. They were not, in fact,
telling me to create an empty div to "get things to line up
correctly"... Whoops.

Note #2: This is not a proper commit message, this is a note to myself
as this is a learning exercise, if this point ever becomes relevant.